refactor: move query uitl functions to helpers.ts
